+++ changelog.md
@@ -2,3 +2,4 @@
 
 * Implemented filtering considered files by extension, inclusion and exclusion
 * Implemented bytewise file comparisons using mmaps. It's faster than I expected
+* Use fmmap to read full hashes